refactor: move fetch-related functions from globalApi.svelte.ts to fetch.ts
