+++ src/lib/SideBars/DevTool.svelte
@@ -3,7 +3,7 @@
     import TextInput from "../UI/GUI/TextInput.svelte";
     import NumberInput from "../UI/GUI/NumberInput.svelte";
     import Button from "../UI/GUI/Button.svelte";
-    import { getRequestLog } from "src/ts/globalApi.svelte";
+    import { getRequestLog } from "src/ts/fetch";
     import { alertMd, alertWait } from "src/ts/alert";
     import Accordion from "../UI/Accordion.svelte";
     import { getCharToken, getChatToken } from "src/ts/tokenizer";
