+++ src/ts/process/mcp/googlesearchclient.ts
@@ -1,6 +1,6 @@
 import { MCPClientLike } from "./internalmcp";
 import type { MCPTool, RPCToolCallContent } from "./mcplib";
-import { fetchNative } from "../../globalApi.svelte";
+import { fetchNative } from "../../fetch";
 import { alertInput } from "../../alert";
 import localforage from "localforage";
