+++ src/ts/process/memory/hypamemoryv2.ts
@@ -2,7 +2,7 @@ import localforage from "localforage";
 import { type HypaModel, localModels } from "./hypamemory";
 import { TaskRateLimiter, TaskCanceledError } from "./taskRateLimiter";
 import { runEmbedding } from "../transformers";
-import { globalFetch } from "src/ts/globalApi.svelte";
+import { globalFetch } from "src/ts/fetch";
 import { getDatabase } from "src/ts/storage/database.svelte";
 import { appendLastPath } from "src/ts/util";
 import { isMobile } from "src/ts/platform";
