+++ src/ts/process/memory/hypav2.ts
@@ -8,7 +8,7 @@ import type { OpenAIChat } from "../index.svelte";
 import type { ChatTokenizer } from "src/ts/tokenizer";
 import { requestChatData } from "../request/request";
 import { HypaProcesser } from "./hypamemory";
-import { globalFetch } from "src/ts/globalApi.svelte";
+import { globalFetch } from "src/ts/fetch";
 import { runSummarizer } from "../transformers";
 import { parseChatML } from "src/ts/parser.svelte";
