+++ src/ts/process/memory/supaMemory.ts
@@ -7,7 +7,7 @@ import { tokenize, type ChatTokenizer } from "../../tokenizer";
 import { requestChatData } from "../request/request";
 import { HypaProcesser } from "./hypamemory";
 import { stringlizeChat } from "../stringlize";
-import { globalFetch } from "src/ts/globalApi.svelte";
+import { globalFetch } from "src/ts/fetch";
 import { runSummarizer } from "../transformers";
 import { getUserName } from "src/ts/persona";
 import { parseChatML } from "src/ts/parser.svelte";
