+++ src/ts/process/models/nai.ts
@@ -1,6 +1,6 @@
 import { getDatabase, setDatabase } from "src/ts/storage/database.svelte"
 import type { OpenAIChat } from "../index.svelte"
-import { globalFetch } from "src/ts/globalApi.svelte"
+import { globalFetch } from "src/ts/fetch"
 import { alertError, alertInput, alertNormal, alertWait } from "src/ts/alert"
 import { sleep } from "src/ts/util"
 import { getUserName } from "src/ts/persona"
