+++ src/ts/process/scriptings.ts
@@ -18,7 +18,8 @@ import { v4 } from "uuid";
 import { getModuleLorebooks, getModuleTriggers } from "./modules";
 import { Mutex } from "../mutex";
 import { tokenize } from "../tokenizer";
-import { fetchNative, readImage } from "../globalApi.svelte";
+import { readImage } from "../globalApi.svelte";
+import { fetchNative } from "../fetch";
 import { loadLoreBookV3Prompt } from './lorebook.svelte';
 import { getPersonaPrompt } from "../persona";
 import { getUserIcon } from "../persona";
