+++ src/ts/process/stableDiff.ts
@@ -3,7 +3,8 @@ import { getDatabase } from "../storage/database.svelte"
 import { type character } from '../storage/types/character'
 import { requestChatData } from "./request/request"
 import { alertError } from "../alert"
-import { fetchNative, globalFetch, readImage } from "../globalApi.svelte"
+import { readImage } from "../globalApi.svelte"
+import { globalFetch, fetchNative } from "../fetch"
 import { CharEmotion } from "../stores.svelte"
 import type { OpenAIChat } from "./index.svelte"
 import { processZip } from "./processzip"
