+++ src/ts/process/tts.ts
@@ -2,7 +2,8 @@ import { alertError } from "../alert";
 import { getCurrentCharacter, getDatabase } from "../storage/database.svelte";
 import { type character } from '../storage/types/character';
 import { runTranslator, translateVox } from "../translator/translator";
-import { globalFetch, loadAsset } from "../globalApi.svelte";
+import { loadAsset } from "../globalApi.svelte";
+import { globalFetch } from "../fetch";
 import { language } from "src/lang";
 import { sleep } from "../util";
 import { runVITS } from "./transformers";
